extracting InMemoryEventStoreFactory functionality into AbstractInMemoryEventStoreFactory
